+++ cms/src/components/content/content-block.json
@@ -23,6 +23,9 @@
         "videos",
         "audios"
       ]
+    },
+    "bodyMarkdown": {
+      "type": "richtext"
     }
   }
 }
