+++ cms/src/extensions/documentation/documentation/1.0.0/full_documentation.json
@@ -14,7 +14,7 @@
       "name": "Apache 2.0",
       "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
     },
-    "x-generation-date": "2025-04-27T23:32:15.021Z"
+    "x-generation-date": "2025-04-27T23:40:01.498Z"
   },
   "x-strapi-config": {
     "plugins": [
@@ -5501,6 +5501,9 @@
                 }
               }
             }
+          },
+          "bodyMarkdown": {
+            "type": "string"
           }
         }
       },
